requirements.txt
requests
pandas
pandas_ta
logzero
python-dotenv
smartapi-python
pyotp
